+++ src/main/java/co/uk/jiveelection/campaign/App.java
@@ -1,7 +1,7 @@
 package co.uk.jiveelection.campaign;
 
 import co.uk.jiveelection.campaign.jive.memory.InMemoryJiveTranslator;
-import co.uk.jiveelection.campaign.output.twit.TwitterOutput;
+import co.uk.jiveelection.campaign.output.twitter.TwitterOutput;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import twitter4j.TwitterException;
